+++ Assets/Scripts/DesktopUI/EditControllerBindingsButton.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+
+namespace EVRC.DesktopUI
+{
+    /**
+     * Behaviour used on a UI Button to open the Controller Bindings interface
+     */
+    public class EditControllerBindingsButton : MonoBehaviour
+    {
+        public void Activate()
+        {
+            ActionsControllerBindingsLoader.CurrentBindingsController?.EditBindings();
+        }
+    }
+}